restructuring the backend to be a real python package. not sure if this is really all that important.
